+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## [0.11.0-beta] - 2025-04-08
+## [0.11.0-beta] - 2025-04-11
 
 **Changed:**
 
-- *#46* If the `name` of a `TrainerNPC` is unset or empty it will now fall back to the display name of the entity the trainer is attached to
+- *#49* Added `nickname` property to `PokemonModel`s (TODO)
+- *#48* Adjustments to `RCTBattleAI` *move* and *switch* evaluations (e.g. better awareness of shedinjas wonderguard) and other tweaks
+- *#47* If the `name` of a `TrainerNPC` is unset or empty it will now fall back to the display name of the entity the trainer is attached to
+
+**Fixed:**
+
+- *#46* `RCTBattleAI` attempting to choose illegal actions and causing battles to softlock in certain situations (e.g. switching out against pokemon with "arena trap" or while "mean look" is active)
 
 **Removed:**